chore(core): improve type hints and simplify code structure

+++ src/domain/payment/adapters/paypal_adapter.py
@@ -1,4 +1,5 @@
 import uuid
+from decimal import Decimal
 
 from domain.payment.model.value_objects import PaymentId
 from domain.payment.ports.payment_adapter_interface import PaymentAdapterInterface
@@ -7,7 +8,7 @@ from domain.payment.ports.payment_adapter_interface import PaymentAdapterInterfa
 class PayPalPaymentAdapter(PaymentAdapterInterface):
     """Mock PayPal adapter for demonstration purposes."""
 
-    async def new_payment(self, total_price: float) -> PaymentId:
+    async def new_payment(self, total_price: Decimal) -> PaymentId:
         """Simulate creation of a new PayPal payment."""
         return PaymentId(str(uuid.uuid4()))
